fix: expired jwt token should redirect (ENG-2855)
Environments with multiple keys defined on a JWT token (fairly common)
will end up with the keys not used to sign the token cause an
InvalidTokenError, which we don't want to redirect to login on to avoid
loops, even if another key was used and could be verified. This came up
with expired tokens not properly redirecting to the login page, but it
could also happen with entirely valid tokens depending on the order of
the keys return by the key server.

Fixes: [ENG-2855](https://stacklet.atlassian.net/browse/ENG-2855)